adicionar estruturas lógicas e refatorar parte do código

+++ sinal.py
@@ -6,17 +6,25 @@ from time import sleep
 def calcular_sinal(distancia: float) -> float:
     return max(((distancia ** 2) * -(16 / 900)) + 100, 0)
 
+# cria uma recomendação para o usuário a partir da intensidade do sinal em %
 def recomendar_acao(intensidade: float) -> str:
     if intensidade < 30:
         return "Tente adicionar um repetidor de sinal Wi-Fi em sua casa"
     if intensidade < 50:
         return "Tente mover o roteador para outro cômodo"
-    if intensidade < 75:
+    if intensidade < 80:
         return "A intensidade de sinal será boa"
     return "A intensidade de sinal será muito boa"
 
+def imprimir_dados(comodo: str, intensidade: float):
+    acao = recomendar_acao(intensidade)
+    print(f"O cômodo {comodo} terá sinal com intensidade de {intensidade:.1f}%")
+    print(acao)
+    print("-" * len(acao))
+
 dados = pd.DataFrame(columns=["cômodo", "intensidade"])
 
+# carregar os dados do arquivo se ele existir
 if os.path.isfile("comodos.csv"):
     dados = pd.read_csv("comodos.csv")
 
@@ -36,6 +44,9 @@ while True:
             nome = input("Nome do cômodo: ")
             dist = float(input("Distância do roteador (em m): "))
             intensidade = calcular_sinal(dist)
+
+            imprimir_dados(nome, intensidade)
+
             novo = pd.DataFrame({"cômodo": [nome], "intensidade": [intensidade]})
             # pandas retorna um erro se concatenarmos um DataFrame a um DataFrame vazio
             if len(dados) > 0:
@@ -44,17 +55,16 @@ while True:
                 dados = novo
         case "2":
             for _, row in dados.iterrows():
-                acao = recomendar_acao(float(row["intensidade"]))
-                print(f"O cômodo {row["cômodo"]} terá sinal com intensidade de {float(row["intensidade"]):.1f}%")
-                print(acao)
-                print("-" * len(acao))
-                sleep(1)
+                # checa se os valores existem na linha atual
+                if row["cômodo"] is not None and row["intensidade"] is not None:
+                    imprimir_dados(row["cômodo"], float(row["intensidade"]))
+                    sleep(1)
         case "3":
             dados = pd.DataFrame(columns=dados.columns)
         case "4":
             if len(dados) > 0:
                 dados.to_csv("comodos.csv")
-            elif os.path.isfile("comodos.csv"):
+            elif os.path.isfile("comodos.csv") and len(dados) == 0:
                 os.remove("comodos.csv")
             break
         case "5":
